// idcs_portal/src/store/models/Product/Product.ts
// INTEL CONFIDENTIAL
// Copyright (C) 2023 Intel Corporation

export default interface Product {
  id: string
  familyId: string
  vendorId: string
  created: string
  name: string
  description: string
  category: string
  recommendedUseCase: string
  cpuSockets: string
  cpuCores: string
  diskSize: string
  displayName: string
  familyDisplayDescription: string
  familyDisplayName: string
  information: string
  instanceType: string
  instanceCategories: string
  memorySize: string
  nodesCount: string
  processor: string
  region: string
  service: string
  eccn: string
  pcq: string
  accountType: string
  unit: string
  rate: string
  usageExpr: string
  imageSource: string
  releaseStatus: string
  useCase?: string
  duration?: string
  displayOrder?: string
  waitlist?: boolean
  extraTextInput1Label?: string
  extraTextInput1Type?: string
  extraTextInput1Optional?: boolean
  extraTextInput2Label?: string
  extraTextInput2Type?: string
  extraTextInput2Optional?: boolean
  extraTextInput3Label?: string
  extraTextInput3Type?: string
  extraTextInput3Optional?: boolean
  extraTextInput4Label?: string
  extraTextInput4Type?: string
  extraTextInput4Optional?: boolean
  extraTextInput5Label?: string
  extraTextInput5Type?: string
  extraTextInput5Optional?: boolean
  extraDropdown1Options?: string[]
  extraDropdown1Label?: string
  extraDropdown2Options?: string[]
  extraDropdown2Label?: string
  extraDropdown3Options?: string[]
  extraDropdown3Label?: string
  extraDropdown4Options?: string[]
  extraDropdown4Label?: string
  extraDropdown5Options?: string[]
  extraDropdown5Label?: string
}